fix the merge of 4a63850fe1607bedb78aa55ef870234f5ae9aa84
rebase #16

to the master(after moving to pluggable sources)
possibly fixes #26

+++ src/main/scala/com/scylladb/migrator/readers/Cassandra.scala
@@ -106,7 +106,7 @@ object Cassandra {
                 if (row.isNullAt(ttlOrdinal)) None
                 else Some(row.getInt(ttlOrdinal)),
                 if (row.isNullAt(writetimeOrdinal)) None
-                else Some(row.getInt(writetimeOrdinal)))
+                else Some(row.getLong(writetimeOrdinal)))
           }
           .groupBy {
             case (fieldName, value, ttl, writetime) => (ttl, writetime)
